requirement.txt
# Core ML
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
bitsandbytes>=0.41.0   # For 4-bit/8-bit quantization
scikit-learn>=1.0.0
numpy>=1.20.0
pandas>=2.0.0

# Utilities
tqdm
python-dotenv
pyyaml
jinja2
scipy

# APIs (Optional but recommended)
openai>=1.0.0
google-genai>=0.1.0    # Check specific version for Gemini